Implemented sprint 2 stuff to work with new level/room setup
