+++ Team4-LegendOfZelda.csproj
@@ -52,6 +52,7 @@
     <Compile Include="ICommand Classes\Commands\LinkUseBoomerangCommand.cs" />
     <Compile Include="ILevel Classes\IRoom.cs" />
     <Compile Include="ILevel Classes\Levels\Level1.cs" />
+    <Compile Include="ILevel Classes\Levels\Sprint2Level.cs" />
     <Compile Include="IPlayer Classes\Link Classes\Link States\LinkKnockbackEastState.cs" />
     <Compile Include="IPlayer Classes\Link Classes\Link States\LinkKnockbackNorthState.cs" />
     <Compile Include="IPlayer Classes\Link Classes\Link States\LinkKnockbackSouthState.cs" />
@@ -213,7 +214,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ICommand Classes\Commands\QuitCommand.cs" />
     <Compile Include="ISprite Classes\Sprite\TextSprite.cs" />
-    <Compile Include="ILevel Classes\Levels\Sprint2Level.cs" />
+    <Compile Include="ILevel Classes\Levels\Sprint2Room.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="MonoGame.Framework">
